requirements.txt
# Core Framework
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0

# Computer Vision & Image Processing
opencv-python>=4.9.0
Pillow>=10.0.0
scikit-image>=0.24.0
scikit-learn>=1.5.0
pycocotools>=2.0.0
matplotlib>=3.9.0
imageio>=2.30.0

# 3D Processing
open3d>=0.18.0
trimesh>=4.4.0

# Deep Learning Models
transformers>=4.45.0
diffusers>=0.31.0
peft>=0.12.0
bitsandbytes>=0.43.0
timm>=1.0.0
einops>=0.8.0
accelerate>=0.34.0
xformers>=0.0.18
sentencepiece>=0.2.0
protobuf>=3.20.0

# Data & HuggingFace Ecosystem
datasets>=2.20.0
huggingface_hub>=0.25.0
packaging>=24.0

# API & LLM
openai>=1.50.0

# Configuration & Utilities
omegaconf>=2.3.0
hydra-core>=1.3.0
tqdm>=4.66.0
requests>=2.31.0
easydict>=1.13
scipy>=1.13.0
wandb>=0.18.0
pyyaml>=6.0.0
termcolor>=2.0.0
networkx>=3.0.0
rich>=13.0.0
typing_extensions>=4.0.0

# GigaPose & Grounding DINO dependencies
# These often require manual installation or cloning
# -e git+https://github.com/thodan/bop_toolkit.git@e7ba9f2389d48d721856df8147fb133c0145e99c#egg=bop_toolkit_lib
# megapose (custom)
# gdino (custom)
